refactor(e2e): rename guest client and move CCM tests to own file
- Rename GetGuestClient() to GetHostedClusterClient() to align with
  v2 framework terminology (hosted cluster, not guest cluster)
- Move GCPCloudControllerManagerTest from control_plane_workloads_test.go
  to hosted_cluster_ccm_test.go with RegisterHostedClusterCCMTests
  registration pattern, separating hosted cluster validation from
  management cluster workload tests
